requirements.txt
PyYAML==3.11
requests==2.3.0
wsgiref==0.1.2
